@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-gd": "*",
         "algolia/algoliasearch-client-php": "^3.3",
         "algolia/scout-extended": "^3.0",
@@ -44,7 +44,7 @@
         "raiym/instagram-php-scraper": "^0.12.9",
         "spatie/eloquent-sortable": "^4.0",
         "spatie/enum": "^3.13",
-        "spatie/laravel-data": "^3.6",
+        "spatie/laravel-data": "^4",
         "spatie/laravel-event-sourcing": "^7.3",
         "spatie/laravel-ignition": "^2.4",
         "spatie/laravel-model-states": "^2.4",
@@ -68,7 +68,7 @@
         "eventsauce/test-utilities": "^3.4",
         "fakerphp/faker": "^1.9.1",
         "laravel/breeze": "^1.21",
-        "laravel/dusk": "^7.11",
+        "laravel/dusk": "^8.2",
         "laravel/pint": "^1.0",
         "laravel/sail": "^1.18",
         "mockery/mockery": "^1.4.4",
@@ -88,7 +88,7 @@
             "Domain\\Profiles\\": "src/Domains/Profiles",
             "Domain\\Work\\": "src/Domains/Work",
             "Domain\\Work2\\": "src/Domains/Work2/src",
-            "Domain\\Work2\\Tests\\": "src/Domains/Work2/tests",
+            "Tests\\Work2\\": "src/Domains/Work2/tests",
 
             "Support\\": "src/Support/",
             "Database\\Factories\\": "database/factories/",
