@@ -4,6 +4,7 @@ import {HTMLAttributes, ReactElement} from "react";
 import { Method } from '@inertiajs/core';
 
 type Props = {
+    id?: string;
     href: string;
     method?: Method;
     title?: ReactElement|string;
@@ -11,7 +12,7 @@ type Props = {
     icon?: ReactElement|boolean;
 }
 
-export function CtaLink({href, variant="primary", method="get", title, className='', icon=<RightArrow />, children}: Props & HTMLAttributes<HTMLDivElement>) {
+export function CtaLink({id, href, variant="primary", method="get", title, className='', icon=<RightArrow />, children}: Props & HTMLAttributes<HTMLDivElement>) {
 
     const backgroundColor = {
         primary: 'bg-teal text-white',
@@ -20,7 +21,7 @@ export function CtaLink({href, variant="primary", method="get", title, className
 
 
     return (
-        <Link method={method} href={href} className={`${className} ${backgroundColor[variant]} justify-between font-semibold rounded-lg flex text-center p-4`}>
+        <Link id={id} method={method} href={href} className={`${className} ${backgroundColor[variant]} justify-between font-semibold rounded-lg flex text-center p-4`}>
             <span className={"hidden"}>&nbsp;</span>
             {children || title}
             {icon ? icon : <span>&nbsp;</span>}
