@@ -10,28 +10,17 @@ use Domain\Profiles\Repositories\PhotoRepository;
 use Domain\Profiles\Repositories\VideoRepository;
 use Domain\Work\Models\Application;
 use Domain\Work2\Actions\Apply;
+use Domain\Work2\Actions\ExtendApplication;
 use Domain\Work2\Data\ApplyData;
+use Domain\Work2\Models\Listing;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class ApplicationController extends Controller
 {
-    public function index()
-    {
-        return redirect()->to("/dashboard");
-    }
-
-    public function show(Application $application)
-    {
-        $role = $application->role;
-
-        return Inertia::render('Roles/Show')
-            ->with("viewModel", new ModelRoleViewModel($role));
-    }
-
     public function create(Role $role)
     {
-        return Inertia::render('Applications/Create')
+        return Inertia::render('Roles/Listings/Apply')
             ->with("viewModel", new ModelRoleViewModel($role))
             ->with("meViewModel", new ModelMeViewModel(
                 auth()->user()
@@ -39,37 +28,31 @@ class ApplicationController extends Controller
             ));
     }
 
-
     public function store(Role $role, Request $request)
     {
-        $this->validate($request, [
-            'cover_letter' => 'required',
-            'digitals' => 'required',
-            'photos' => 'required',
-            'available_dates' => 'required',
-            'brand_conflicted' => 'required',
-        ]);
-
         $model = auth()->user();
         if (!($model instanceof Model)) abort(403, "You are not a model");
 
-        $data = ApplyData::fromRequest($model, $request->all());
+        $data = ApplyData::fromRequest($request->all());
+
         app()->make(Apply::class)($model, $role, $data);
 
         return redirect()->route("roles.show", $role);
     }
 
-    public function update(Application $application)
+    public function update(Role $role, Request $request)
     {
-        if ($videos = request()->get("casting_videos")) {
-            app(VideoRepository::class)->update($application, Application::CASTING_VIDEOS, $videos);
-        }
+        $listing = Listing::where("role_id", $role->id)
+            ->where("model_id", auth()->id())
+            ->first();
 
-        if ($photos = request()->get("casting_photos")) {
-            app(PhotoRepository::class)->update($application, Application::CASTING_PHOTO_FOLDER, $photos);
-        }
+        app()->make(ExtendApplication::class)(
+            $listing,
+            $request->get("casting_photos"),
+            $request->get("casting_videos")
+        );
 
-        return Inertia::render('Applications/Updated')
-            ->with("viewModel", new ModelRoleViewModel($application->role));
+        return Inertia::render('Roles/Listings/Updated')
+            ->with("viewModel", new ModelRoleViewModel($listing->role, $listing));
     }
 }
