@@ -2,9 +2,8 @@
 
 namespace App\Http\Controllers;
 
-use Domain\Jobs\Data\ApplyData;
 use Domain\Jobs\Models\Role;
-use Domain\Work\Actions\Pass;
+use Domain\Work2\Actions\Pass;
 
 class PassController extends Controller
 {
