@@ -9,17 +9,18 @@ use Inertia\Inertia;
 
 class RoleController extends Controller
 {
+    public function index()
+    {
+        return redirect()->to("/dashboard");
+    }
+
     public function show(Role $role)
     {
         Session::put('viewed_roles', array_unique([$role->id, ...Session::get('viewed_roles', [])]));
 
-        $application = $role->applications->firstWhere('model_id', auth()->id());
-
-        if ($application && $application->photos()->count()===0) {
-            return redirect()->route('roles.apply', $role->id);
-        }
+        $listing = $role->listings()->where('model_id', auth()->id())->first();
 
         return Inertia::render('Roles/Show')
-            ->with("viewModel", new ModelRoleViewModel($role));
+            ->with("viewModel", new ModelRoleViewModel($role, $listing));
     }
 }
