@@ -2,21 +2,18 @@
 
 namespace App\Nova\Actions;
 
-use Domain\Jobs\Models\Role;
-use Domain\Work2\Actions\Invite;
 use Illuminate\Bus\Queueable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Collection;
 use Laravel\Nova\Actions\Action;
 use Laravel\Nova\Fields\ActionFields;
-use Laravel\Nova\Fields\Select;
+use Laravel\Nova\Fields\Text;
+use Laravel\Nova\Fields\Textarea;
 use Laravel\Nova\Http\Requests\NovaRequest;
 
-class SendInviteToListing extends Action
+class Hire extends Action
 {
     use InteractsWithQueue, Queueable;
-
-    public $name = 'Invite';
 
     /**
      * Perform the action on the given models.
@@ -25,15 +22,13 @@ class SendInviteToListing extends Action
      * @param  \Illuminate\Support\Collection  $models
      * @return mixed
      */
-    public function handle(ActionFields $fields, Collection $listings)
+    public function handle(ActionFields $fields, Collection $collection)
     {
-        foreach ($listings as $listing) {
-            app(Invite::class)->execute(
-                $listing->role,
-                $listing->model
-            );
-
-            Action::message("Invite for role has been sent to " . $listing->model->name);
+        foreach ($collection as $listing) {
+            app(\Domain\Work2\Actions\Hire::class)
+                ->execute(
+                    $listing,
+                );
         }
     }
 
