@@ -3,7 +3,6 @@
 namespace App\Nova\Actions;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Support\Collection;
 use Laravel\Nova\Actions\Action;
@@ -25,10 +24,10 @@ class Reject extends Action
      */
     public function handle(ActionFields $fields, Collection $collection)
     {
-        foreach ($collection as $application) {
-            app(\Domain\Work\Actions\Reject::class)
+        foreach ($collection as $listing) {
+            app(\Domain\Work2\Actions\Reject::class)
                 ->execute(
-                    $application,
+                    $listing,
                     $fields->subject,
                     $fields->message
                 );
