@@ -3,10 +3,14 @@
 namespace App\Nova;
 
 use App\Nova\Actions\CreatePresentation;
+use App\Nova\Actions\Hire;
+use App\Nova\Actions\Reject;
 use App\Nova\Actions\SendInviteToListing;
+use App\Nova\Actions\Shortlist;
 use Laravel\Nova\Fields\BelongsTo;
 use Laravel\Nova\Fields\Boolean;
 use Laravel\Nova\Fields\Line;
+use Laravel\Nova\Fields\MorphMany;
 use Laravel\Nova\Fields\Stack;
 use Laravel\Nova\Fields\Text;
 use Laravel\Nova\Fields\Textarea;
@@ -16,7 +20,7 @@ class Listing extends Resource
 {
     public static $model = \Domain\Work2\Models\Listing::class;
 
-    
+
     public function title() {
         return $this->model->name .' at '. $this->role->name;
     }
@@ -40,7 +44,7 @@ class Listing extends Resource
     {
         return [
             BelongsTo::make("Model")->searchable(),
-            BelongsTo::make("Role"),
+            BelongsTo::make("Role")->hideFromIndex(),
             Stack::make("Name", [
                 Line::make("Role", function () {
                     return $this->role->name;
@@ -54,24 +58,46 @@ class Listing extends Resource
             ])->onlyOnIndex(),
             Boolean::make("Invited", "invited")
                 ->readonly()
-                ->onlyOnIndex()
                 ->filterable(fn ($request, $query, $value, $attributes) => $query->{$value ? "whereNotNull" : "whereNull"}('invited_at') )
                 ->resolveUsing(fn($request, $model) => $model['invited_at']!==null),
             Boolean::make("Applied", "applied")
                 ->readonly()
-                ->onlyOnIndex()
                 ->filterable(fn ($request, $query, $value, $attributes) => $query->{$value ? "whereNotNull" : "whereNull"}('applied_at') )
                 ->resolveUsing(fn($request, $model) => $model['applied_at']!==null),
-            Text::make('Photos', fn() =>
-                '<div class="flex overflow-x-scroll overflow-y-hidden" style="height: 120px;">' .
-                        $this->photos->map(fn($photo) => '<img src="' . $photo->cdn_path_thumb . '" />')->implode('')
-                . '</div>'
-            )->asHtml(),
-            Text::make('Digitals', fn() =>
-                '<div class="flex overflow-x-scroll overflow-y-hidden" style="height: 120px;">' .
-                        $this->digitals->map(fn($photo) => '<img src="' . $photo->cdn_path_thumb . '" />')->implode('')
-                . '</div>'
-            )->asHtml(),
+
+            Boolean::make("Shortlisted", "shortlisted")
+                ->readonly()
+                ->filterable(fn ($request, $query, $value, $attributes) => $query->{$value ? "whereNotNull" : "whereNull"}('shortlisted_at') )
+                ->resolveUsing(fn($request, $model) => $model['shortlisted_at']!==null),
+
+            Boolean::make("Hired", "hired")
+                ->readonly()
+                ->filterable(fn ($request, $query, $value, $attributes) => $query->{$value ? "whereNotNull" : "whereNull"}('hired_at') )
+                ->resolveUsing(fn($request, $model) => $model['hired_at']!==null),
+
+            Boolean::make("Rejected")
+                ->readonly()
+                ->filterable(fn ($request, $query, $value, $attributes) => $query->{$value ? "whereNotNull" : "whereNull"}('rejected_at') )
+                ->resolveUsing(fn($request, $model) => $model['rejected_at']!==null),
+
+            Text::make('Photos', fn() => '<div style="display: flex; width: 340px; overflow-x: scroll">
+                        ' .implode("", $this->photos->map(fn($photo) => '<img src="'.$photo->cdn_path_thumb.'" style="height: 120px;" />')->toArray())
+                . '</div>')
+                ->asHtml()->onlyOnIndex(),
+            Text::make('Casting photos', fn() => '<div style="display: flex; width: 340px; overflow-x: scroll">
+                        ' .implode("", $this->casting_photos->map(fn($photo) => '<img src="'.$photo->cdn_path_thumb.'" style="height: 120px;" />')->toArray())
+                . '</div>')
+                ->asHtml()->hideFromIndex(),
+            Text::make('Casting videos', fn() => '<div style="display: flex; width: 140px; overflow-x: scroll">
+                        ' .$this->casting_videos->map(fn($video) => '<video src="'.$video->cdn_path.'" style="height: 120px;" controls="true" />')->implode('')
+                . '</div>')
+                ->asHtml()->hideFromIndex(),
+
+            MorphMany::make("Photos")->showOnIndex(true),
+            MorphMany::make("Casting photos", "casting_photos", Photo::class)->showOnIndex(true),
+            MorphMany::make("Casting videos", "casting_videos", Video::class)->showOnIndex(true),
+
+
             Textarea::make('Cover letter'),
             Textarea::make('Brand conflicted'),
             Text::make('Available dates', fn() => implode(", ", $this->available_dates ?? [])),
@@ -121,7 +147,10 @@ class Listing extends Resource
     {
         return [
             new CreatePresentation(),
-            new SendInviteToListing()
+            new SendInviteToListing(),
+            new Shortlist(),
+            new Reject(),
+            new Hire(),
         ];
     }
 }
