@@ -4,43 +4,32 @@ namespace App\ViewModels;
 
 use DateInterval;
 use DatePeriod;
-use Domain\Jobs\Data\ApplicationData;
+use Domain\Jobs\Data\ListingData;
 use Domain\Jobs\Data\RoleData;
 use Domain\Jobs\Models\Role;
+use Domain\Work2\Models\Listing;
 use Spatie\ViewModels\ViewModel;
 
 /** @typescript */
 class ModelRoleViewModel extends ViewModel
 {
     public RoleData $role;
-    public bool $isInvited;
-    public bool $hasApplied;
-    public bool $hasPassed;
-    public bool $isHired;
+    public ?ListingData $listing = null;
+
+    /** @var array<string> */
     public array $shootDates;
-    public string $status;
-    public ?ApplicationData $my_application;
 
-    public function __construct(Role $role)
+    public bool $hasApplied;
+    public bool $isHired;
+
+    public function __construct(Role $role, Listing $listing = null)
     {
-        $role->load(
-            'my_application',
-            'my_application.hire',
-            'my_application.photos',
-            'my_application.model',
-            'my_application.casting_photos',
-            'my_application.casting_videos',
-            'photos',
-            'public_photos',
-            'job',
-            'job.look_and_feel_photos',
-            'job.brand',
-            'job.client',
-            'my_passes',
-            'my_invites',
-        );
+        $this->role = RoleData::from($role->load("job"));
 
-        $this->role = RoleData::from($role);
+        if ($listing) {
+            $listing->load(["casting_photos", "casting_videos"]);
+            $this->listing = ListingData::from($listing);
+        }
 
         $period = new DatePeriod(
             $role->start_date,
@@ -54,26 +43,7 @@ class ModelRoleViewModel extends ViewModel
             $this->shootDates[] = $date->format('Y-m-d');
         }
 
-        $this->isInvited = $role->my_invites->count() > 0;
-        $this->hasApplied = !!$role->my_application;
-
-        $this->my_application = ($application = $role->my_application)
-            ? ApplicationData::from($application)
-            : null;
-
-        $this->isHired = $role->my_applications->count()>0 && $role->my_applications->first()->hire()->count()>0;
-        $this->hasPassed = !$this->hasApplied && $role->my_passes()->count() > 0;
-
-        if ($this->isHired) {
-            $this->status = "hired";
-        } elseif ($this->hasPassed) {
-            $this->status = "passed";
-        } elseif ($this->isInvited) {
-            $this->status = "invited";
-        } elseif ($this->hasApplied) {
-            $this->status = "applied";
-        } else {
-            $this->status = "open";
-        }
+        $this->hasApplied = !!optional($listing)->applied_at;
+        $this->isHired = !!optional($listing)->hired_at;
     }
 }
