@@ -6,6 +6,9 @@ use Spatie\LaravelData\Data;
 
 /** @typescript */
 class ExtraFieldsData extends Data {
-    public bool $casting_photos;
-    public bool $casting_videos;
+
+    public function __construct(
+        public bool $casting_photos, public  bool $casting_videos)
+    {
+    }
 }
