@@ -2,12 +2,12 @@
 
 namespace Domain\Work2\Policies;
 
-use Domain\Profiles\Models\Model;
+use Domain\Work2\Models\Listing;
 use Illuminate\Contracts\Auth\Authenticatable;
 
 class ListingPolicy
 {
-    public function view(Authenticatable $authenticatable, Model $model)
+    public function view(Authenticatable $authenticatable, Listing $listing)
     {
         return true;
     }
@@ -17,17 +17,17 @@ class ListingPolicy
         return true;
     }
 
-    public function update(Authenticatable $authenticatable, Model $model)
+    public function update(Authenticatable $authenticatable, Listing $listing)
     {
         return true;
     }
 
-    public function addPhotos(Authenticatable $authenticatable, Model $model)
+    public function addPhotos(Authenticatable $authenticatable, Listing $listing)
     {
         return true;
     }
 
-    public function addPhoto(Authenticatable $authenticatable, Model $model)
+    public function addPhoto(Authenticatable $authenticatable, Listing $listing)
     {
         return true;
     }
@@ -40,10 +40,5 @@ class ListingPolicy
     public function viewNova(Authenticatable $authenticatable)
     {
         return false;
-    }
-
-    public function uploadFiles(Model $model): bool
-    {
-        return true;
     }
 }
