@@ -4,11 +4,11 @@ namespace Domain\Work2\Tests\Data;
 
 use Domain\Profiles\Models\Model;
 use Domain\Work2\Data\ApplyData;
-use Domain\Work2\Tests\Mock\AnalysePhotoMock;
-use Domain\Work2\Tests\Mock\MovePhotoMock;
-use Domain\Work2\Tests\Mock\PhashPhotoMock;
 use Illuminate\Support\Facades\Storage;
 use Tests\TestCase;
+use Tests\Work2\Mock\AnalysePhotoMock;
+use Tests\Work2\Mock\MovePhotoMock;
+use Tests\Work2\Mock\PhashPhotoMock;
 
 class ApplyDataTest extends TestCase
 {
