@@ -1,7 +1,6 @@
 <?php
 
-namespace Domain\Work2\Tests\Mock;
-
+namespace Tests\Work2\Mock;
 
 use Support\Actions\MovePhoto;
 
